refactor: 녹화 판단 알고리즘 리팩토링
