+++ package.json
@@ -11,7 +11,6 @@
         "amazon-kinesis-video-streams-webrtc": "^2.0.0",
         "bootstrap": "^5.2.2",
         "dotenv": "^16.0.2",
-        "imagediff": "^1.0.8",
         "react": "^18.2.0",
         "react-bootstrap": "^2.5.0",
         "react-burger-menu": "^3.0.8",
